Allow assets to be concatenated unprocessed to generate a digest of their sources.
This lets us determine if an asset's sources have changed.
If not, then we don't need to recompile and recompress that asset.
If :process => false is passed as an option, the only processors
that are run are Sprockets Directives and ERB.
Other processors, such as coffeescript and compression, are removed.