Allow Setting Keep via Env Var
Right now the keep value in sprockets is a magic number of 2. This PR moves that number to a constant that can be used across multiple files and modified globally. This PR also allows the manipulation of this value from the command line or other processes via an environment variable. This would be extremely useful for running command from terminal. For example to run clean and keep only one backup you could run a rake task like this:

```rub
SPROCKETS_CLEAN_KEEP=1 rake assets:clean
```

Assuming the user was also using the sprockets-rails tasks.