Improve functionality of Manifest path and dir to accommodate "directory-only" path.
Previously, the Manifest path parameter had to be a full filename. The manifest-xxx.json file path would only be generated if path itself was blank. In some cases, this would cause an error at the time Manifest.save was called.

This commits adds several test cases which illustrate the change in functionality. The previous testcases seem to imply that this functionality should have existed, but was not covered.